requirements.txt
google-cloud-aiplatform
google-adk
flask
gunicorn
functions-framework